frontend: Break up dataService as it is getting big.
